Moved around some code
